# More lookaside cache optimizations