--- a/enterprise/server/backends/distributed/distributed.go
+++ b/enterprise/server/backends/distributed/distributed.go
@@ -344,6 +344,7 @@ func (t *teeReadCloser) Read(p []byte) (int, error) {
 	return read, t.lastReadErr
 }
 
+
 func (t *teeReadCloser) Close() error {
 	err := t.rc.Close()
 	if err == nil && t.lastReadErr != nil && t.lastReadErr != io.EOF {
@@ -407,6 +408,20 @@ func (c *Cache) setLookasideEntry(lookasideKey string, data []byte) {
 	c.log.Debugf("Set %q in lookaside cache", lookasideKey)
 }
 
+var lookasideCacheLookupCount map[bool]prometheus.Counter
+
+func init() {
+	// Calling LookasideCacheLookupCount.With is a large portion of the time
+	// spent on the lookaside cache, so just do it once.
+	lookasideCacheLookupCount = make(map[bool]prometheus.Counter, 2)
+	lookasideCacheLookupCount[true] = metrics.LookasideCacheLookupCount.With(prometheus.Labels{
+		metrics.LookasideCacheLookupStatus: metrics.HitStatusLabel,
+	})
+	lookasideCacheLookupCount[false] = metrics.LookasideCacheLookupCount.With(prometheus.Labels{
+		metrics.LookasideCacheLookupStatus: metrics.MissStatusLabel,
+	})
+}
+
 // getLookasideEntry returns the resource and if it was found in the lookaside
 // cache.
 func (c *Cache) getLookasideEntry(r *rspb.ResourceName) ([]byte, bool) {
@@ -432,14 +447,7 @@ func (c *Cache) getLookasideEntry(r *rspb.ResourceName) ([]byte, bool) {
 	}
 	c.lookasideMu.Unlock()
 
-	lookupStatus := metrics.MissStatusLabel
-	if found {
-		lookupStatus = metrics.HitStatusLabel
-	}
-	metrics.LookasideCacheLookupCount.With(prometheus.Labels{
-		metrics.LookasideCacheLookupStatus: lookupStatus,
-	}).Inc()
-
+	lookasideCacheLookupCount[found].Inc()
 	if found {
 		c.log.Debugf("Got %q from lookaside cache", k)
 		return entry.data, true