--- a/server/remote_cache/digest/digest.go
+++ b/server/remote_cache/digest/digest.go
@@ -217,17 +217,22 @@ func (r *CASResourceName) DownloadString() string {
 	// Normalize slashes, e.g. "//foo/bar//"" becomes "/foo/bar".
 	instanceName := filepath.Join(filepath.SplitList(r.GetInstanceName())...)
 	if isOldStyleDigestFunction(r.rn.DigestFunction) {
-		return fmt.Sprintf(
-			"%s/%s/%s/%d",
-			instanceName, blobTypeSegment(r.GetCompressor()),
-			r.GetDigest().GetHash(), r.GetDigest().GetSizeBytes())
-	} else {
-		return fmt.Sprintf(
-			"%s/%s/%s/%s/%d",
-			instanceName, blobTypeSegment(r.GetCompressor()),
-			strings.ToLower(r.rn.DigestFunction.String()),
-			r.GetDigest().GetHash(), r.GetDigest().GetSizeBytes())
+		return strings.Join(
+			[]string{
+				instanceName,
+				blobTypeSegment(r.GetCompressor()),
+				r.GetDigest().GetHash(),
+				strconv.Itoa(int(r.GetDigest().GetSizeBytes()))},
+			"/")
 	}
+	return strings.Join(
+		[]string{
+			instanceName,
+			blobTypeSegment(r.GetCompressor()),
+			strings.ToLower(r.rn.DigestFunction.String()),
+			r.GetDigest().GetHash(),
+			strconv.Itoa(int(r.GetDigest().GetSizeBytes()))},
+		"/")
 }
 
 // NewUploadString returns a new string representing the resource name for
@@ -235,21 +240,26 @@ func (r *CASResourceName) DownloadString() string {
 func (r *CASResourceName) NewUploadString() string {
 	// Normalize slashes, e.g. "//foo/bar//"" becomes "/foo/bar".
 	instanceName := filepath.Join(filepath.SplitList(r.GetInstanceName())...)
-	u := guuid.New()
+	u := guuid.New().String()
 	if isOldStyleDigestFunction(r.rn.DigestFunction) {
-		return fmt.Sprintf(
-			"%s/uploads/%s/%s/%s/%d",
-			instanceName, u.String(), blobTypeSegment(r.GetCompressor()),
-			r.GetDigest().GetHash(), r.GetDigest().GetSizeBytes(),
-		)
-	} else {
-		return fmt.Sprintf(
-			"%s/uploads/%s/%s/%s/%s/%d",
-			instanceName, u.String(), blobTypeSegment(r.GetCompressor()),
-			strings.ToLower(r.rn.DigestFunction.String()),
-			r.GetDigest().GetHash(), r.GetDigest().GetSizeBytes(),
-		)
+		return strings.Join(
+			[]string{
+				instanceName,
+				"uploads",
+				u,
+				blobTypeSegment(r.GetCompressor()),
+				r.GetDigest().GetHash(),
+				strconv.Itoa(int(r.GetDigest().GetSizeBytes()))}, "/")
 	}
+	return strings.Join(
+		[]string{
+			instanceName,
+			"uploads",
+			u,
+			blobTypeSegment(r.GetCompressor()),
+			strings.ToLower(r.rn.DigestFunction.String()),
+			r.GetDigest().GetHash(),
+			strconv.Itoa(int(r.GetDigest().GetSizeBytes()))}, "/")
 }
 
 type ACResourceName struct {
@@ -262,17 +272,24 @@ func (r *ACResourceName) ActionCacheString() string {
 	// Normalize slashes, e.g. "//foo/bar//"" becomes "/foo/bar".
 	instanceName := filepath.Join(filepath.SplitList(r.GetInstanceName())...)
 	if isOldStyleDigestFunction(r.rn.DigestFunction) {
-		return fmt.Sprintf(
-			"%s/%s/ac/%s/%d",
-			instanceName, blobTypeSegment(r.GetCompressor()),
-			r.GetDigest().GetHash(), r.GetDigest().GetSizeBytes())
-	} else {
-		return fmt.Sprintf(
-			"%s/%s/ac/%s/%s/%d",
-			instanceName, blobTypeSegment(r.GetCompressor()),
-			strings.ToLower(r.rn.DigestFunction.String()),
-			r.GetDigest().GetHash(), r.GetDigest().GetSizeBytes())
+		return strings.Join(
+			[]string{
+				instanceName,
+				blobTypeSegment(r.GetCompressor()),
+				"ac",
+				r.GetDigest().GetHash(),
+				strconv.Itoa(int(r.GetDigest().GetSizeBytes()))},
+			"/")
 	}
+	return strings.Join(
+		[]string{
+			instanceName,
+			blobTypeSegment(r.GetCompressor()),
+			"ac",
+			strings.ToLower(r.rn.DigestFunction.String()),
+			r.GetDigest().GetHash(),
+			strconv.Itoa(int(r.GetDigest().GetSizeBytes()))},
+		"/")
 }
 
 func CacheTypeToPrefix(cacheType rspb.CacheType) string {